wip: Add furhter implementation for mincrcft

+++ hyperv/create-vm.sh
@@ -9,8 +9,13 @@ fi
 mkdir -p /mnt/c/VMs
 mkdir -p /mnt/c/VMs/ISOs
 
+if [ ! -f "/mnt/c/VMs/ISOs/ubuntu-25.10-live-server-amd64.iso" ]; then
+  echo "Missing ISO file: ubuntu-25.10-live-server-amd64.iso"
+  exit 400
+fi
+
 rm -rf seed.iso
-./generate-iso.sh
+./generate-iso.sh || exit 500
 cp seed.iso /mnt/c/VMs/ISOs/seed.iso
 
 cp ./create-vm.ps1 /mnt/c/VMs/create-vm.ps1
